Add files via upload
In this i have analyze a sales dataset set and then created a dynamic colorful dashboard for it using excel i have summarised the data and created slicers for time and category wise sales also the date column in this data was not in correct format so using python I changed into to correct date format. I have also upload a screenshot of the dashboard.